feat: Prefill fields with EXIF metadata when adding
